Refactor: Change structure of type file and change name of useStore -> useDarkModeStore
