+++ src/apis/useContentDelete.ts
@@ -1,0 +1,4 @@
+// 미션: 삭제 구현하기(delete 활용)
+// 삭제 버튼을 눌렀을 때 mutation
+// 삭제 버튼을 눌렀을 때 -> 삭제 버튼 누르고 Link 사용
+// Content.tsx 136번째 줄 onClick
