+++ src/apis/useContentSubmitMutation.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { CardResult } from '../../store/type/card/card';
-import instance from '../instance';
+import { CardResult } from '../store/type/card';
+import instance from './instance';
 
 const useContentSubmitMutation = () => {
   const queryClient = useQueryClient();
